Move schema to separate module
